backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
pydantic>=2.7.0
langchain>=0.1.0
langchain-community>=0.0.38
langchain-text-splitters>=0.0.1
langchain-google-genai>=0.0.6
langchain-chroma>=0.1.0
chromadb>=0.4.22
PyMuPDF>=1.24.0
pdfplumber>=0.10.3
python-dotenv>=1.0.0
google-generativeai>=0.3.2
tiktoken>=0.5.2
